refactor(rpc): make python scripts simpler for now
This commit changes current proof scripts to only use the low-level
class constructors instead of the helper functions/definitions and
comments out some of the helpers which may be changed in the near
term.

The classes themselves seem unlikely to change, but we may review
and modify the helper functions/definitions/etc portion of the API
in the near future.