Add a new experimental "extract_uninterp" command.
This command, when given a term at first-order type, will
symbolically evaluate the term and extract any occurrences
of functions mentioned in the given list as fresh variables.
The arguments passed to the function are collected together into
a tuple and returned together with the fresh variables generated
to stand in for those occurrences.