Add a rough cut at visualizing a proof dependency graph.
We parse a verification summary JSON file and produce
a Graphviz file intended to be fed into `dot` to produce
a visulaization. For large proofs, especially, the layout
is problematic; however, even so, it seems like a helpful
way to scroll through a proof and get a sense of what is
proved and what depends on what.