Bump argo submodule to include docs for method results
This bumps the `argo` submodule to include commit
GaloisInc/argo@b9b3edd,
which adds the ability to document the results of methods (in addition to
their parameters). In addition to dealing with the breaking API changes from
that commit, this adds missing documentation for each method's results.

This also requires bumping the `cryptol` submodule commit to bring in similar
changes introduced in GaloisInc/cryptol#1207.