Introduce `w4_abc_aiger`, similar to old `abc`
This introduces a `w4_abc_aiger` command that invokes ABC as an
external process on a generated AIGER file. Also makes `abc` a synonym
for `w4_abc_aiger`, so that it behaves similarly to how it did with
the `abcBridge`-based `abc` command (specifically, using the same
bit-blasting approach from the `aig` library).

Introducing this command required some refactoring to the types of many
of the prover interfaces, making functions that were in `IO` be in
`TopLevel`, instead. This generally seems to be a simplification.