Tweak the term model to avoid a panic situation.
This would occur when concrete nat values were round-tripped
via a bitvector type in the term model, and into a sequence
indexing primitive.

This tweak fixes a particular instance of this problem, but
highlights the fact that we need much more robust handing of the
sequencing primitives and natural numbers.