Remove a variety of `panic` cases in the saw-core simulator primitives. Also, add a switch that controls if the evalutor will be have in a "symbolic" way.
Adding a new `PrimExcept` constructor that allows primitives to
cause fallback behavior after examining all their arguments.
As with the existing `PrimFilterFun`, this causes term reconstruction
in the term model evaluator, and raises an error in the other evaluators.
We use this new type of primitive to remove cases that would otherwise
panic.

In addition, we add a flag to the `BasePrims` record that controls
if we do symbolic muxing for sequence operations and `ite`. For the
term model, we disable this behavior and fall back earlier on term
reconstruction. This avoids constructing large mux trees in some cases.