Add support for memoized streams to the term model evaluator.
Although it is correct to treat streams according to their
SAWCore datatype definition, the naive implementation as a function
looses sharing that users expect given the behavior of Cryptol
streams.  This patch makes sure to memoize concrete evaluations
of stream values, and blocks evaluation of stream lookups for
values that are not concrete.

This significantly improves the performance of the kinds of
stream programs that arise from Cryptol.