Skip Windows builds in CI for collecting built cabal tests for later use.
The CI process runs the cabal tests in a separate job from the build
process.  To do this, the build process "collects" the tests by
building them and uploading them.  This test-build portion fails for
Windows (symptom is the linker finding a duplicate symbol, cause is
probably misconfiguration and use of multiple versions of the same
dependency).  Since the tests are not run for Windows, this
build+upload/collection step should not be needed for Windows and can
be skipped.