Add a notion of "opaque constants".
These are just `Constant` terms that have no definition.
They are similar to `ExtCns`, except that they are not subject
to substitution.  As a consequence, opaque constants are allowed
to appear in the definition of other constants without being
abstracted over.  Opauqe constants never reduce, and must be
treated as unintepreted in proofs.