Introduce an annotation on sorts that classify types that are expected to be inhabited.
This adds a boolean flag to to the SAWCore AST node for sorts.  For most
purposes (typechecking and such), this annotation is purely advisory
and does not directly affect the SAWCore metatheory.  However, for Coq
export, it is used to identify types arguments to functions that are
required to be inhabited. When appearing in a Pi type, lambda or definition,
such type arguments will be accompanied by an implicit argument
demanding a typeclass instance for the `Inhabited` class.

In concrete SAWCore syntax, this annotaion is attached when using the
new `isort` keyword, which is otherwise identical to `sort`.

This patch only adds the new annotation flag, but does not yet plumb
through the Coq translation changes.