Mergify: switch from strict mode to a merge queue
Previously, we were using Mergify's `smart` mode to serialize the PRs that
Mergify merges. This mode has been deprecated, however, and indeed, there
is a `smart` mode brownout today to force Mergify users to switch over:
https://blog.mergify.com/strict-mode-deprecation/
This PR does just that, based on the migration strategy suggested in that link.