Add the ability to cast pointer setup values in LLVM setup blocks.
This produces a new setup value that points to the same location,
but is assumed to point to values of the provided type. This is useful
when, e.g., verifing code that contains `union` values inside
structs, as the type provided by the bitcode will not accurately
reflect all the intended uses.  It may also be useful for other
"polymorphic" C programming styles.