Fix definition of `ensureTupleType` in SCTypeCheck.hs.
With the new tuple representation, a single tuple type is represented
as not a single AST node, but a whole pile of constructor applications
(`Tuple`, `TypeCons`, `TypeNil`). A single call to WHNF followed by
running the syntactic recognizer is not sufficient; we may need to
evaluate to WHNF separately on each constructor.