Update metadata tracking for verification conditions.
This avoids entering the statements of verification conditions
into the "theorem database", and tracks data about their proofs
separately. This is primarily to avoid retaining the terms representing
these conditions throughout the run of a proof, as the VCs can get
quite large, and are never directly used in the proof of another
theorem.