s2n-tests: Simplify artifact handling
Previously, we uploaded a duplicate SAW binary specifically for use by the
`s2n-tests` job. This is unnecessary, however, as the same binary is contained
within a more general artifact that we upload for the SAW test suite. This
patch removes the duplicate artifact, thereby simplifying the overall CI setup.

Fixes #1726.