Permissions.md: More precise documentation for fieldsh
The documentation for `fieldsh(p)` claims that it acts on a pointer of unknown
size. This isn't really true, however, as it is really a shorthand for
`fieldsh(w,p)`, where `w` is the current pointer size. I found the original
documentation misleading enough to the point where I think it deserves to be
rewritten, which I have accomplished in this patch.

[ci skip]