saw-core-what4: Build muxes of symbolic indexes in big-endian order
SAWCore assumes the invariant that when indexing into an infinite stream using
a symbolic index, the mux tree constructed over the index will test each bit in
big-endian order. This is the responsibility of the `selectV` function found in
`saw-core`, `saw-core-sbv`, and `saw-core-what4`. All of these functions save
for `saw-core-what4`'s `selectV` were upholding this invariant, as
`saw-core-what4`'s was testing each bit in little-endian order, resulting in
the oddities observed in #1703.

This corrects the mistake in `saw-core-what4'`s implementation, fixing #1703.
It also leaves some more documentation to make the fact that each `selectV`
should be proceeding in big-endian order more apparent.