saw-core-coq: Remove silly redefinitions of true/false
We were previously defining the following in the Coq support library:

```coq
Definition true := Datatypes.true.
Definition false := Datatypes.false.
```

This is rather silly, as we could just as well use `Datatypes.true` and
`Datatypes.false` directly. What's more, redefining `true` causes issues in a
subsequent commit that adds `lia` support for `bitvector`s, so it is especially
important to remove `true` for that reason.

This removes the redefined versions of `true` and `false`, and it also tweaks
`SpecialTreatment.hs` to translate SAWCore's `True`/`False` into
`Datatypes.true`/`Datatypes.false`. Although this is technically a breaking
change, it is a change that is very unlikely to break existing proofs (unless
your proofs directly mention `SAWCoreScaffolding.true` or
`SAWCoreScaffolding.false`).

Co-authored-by: Eddy Westbrook <westbrook@galois.com>