saw-core-coq: Deprecate SAWCore compatibility shims in Coq
For historical reasons, the Coq support library has several wrapper functions
like the following:

```coq
Definition Bool := bool.
```

These are extremely silly, however, given that `bool` is already widely used in
the Coq world, and providing our own version of `bool` with different
capitalization causes a proliferation of names that refer to the same thing.
It would be nice to rip out these wrapper functions, but unfortunately, doing
so would induce a fair bit of breakage in downstream code. For now, we simply
record the fact that these functions are deprecated, and we will revisit
whether we want to remove them at a later date.