saw-core-coq: Rudimentary lia support for bitvectors
This adds `SAWCoreBitvectorsZifyU64.v`, a file containing all of the necessary
`Zify`-related instances for `bitvector 64` values needed to allow solving a
large class of `bitvector`-related theorems using Coq's `lia` tactic. This also
includes a small set of examples to ensure that `lia` works as expected.
Although this is far from complete (see the documentation near the top of the
file for a list of limitations), it should be enough to handle the common use
case of `bitvector 64`, which is better than nothing.