Heapster: Avoid shift/reduce conflicts involving bitvector negation
We now ensure that `-` has higher precedence than any other operator using
`happy`'s `%left` pragma. This ensures that `- 42 + 27` parses as `(- 42) +
27`, not `- (42 + 27)` like before.

While I was in town, I added `%expect 0` at the top of the parser to ensure
that `happy` will error out in the future if additional shift/reduce conflicts
are introduced.

Fixes #1794.