Undo `TermFMap` mods and the changes they necessitated
I originally made these changes to avoid computing `TermF` hashes twice, but in
doing so assumed a total absence of term hash collisions. The cost of that
assumption far outweighs the cost of hash recomputation which, at least for
terms with cached subterm hashes (which I expect to encompass most terms), is
actually quite cheap.