Checkpoint: changing abstractions to support customizable script execution caching
The motivation is to let users take advantage of computation they've already
performed when interpreting scripts, when the results of that computation
haven't changed. This provided a good opportunity to split up some of the work
in `SAWT` into lower-level monad manipulations and higher-level statement and
script interpretation utilities.

`SAWT.Interpret` currently exposes both a statement-wise interpretation mode and
a script-wise interpretation mode, with variants for different caching
semantics. Whether these variations are fully compatible with one another
remains to be seen.