Manual: Document pitfalls of specifying code with bools
While the C `bool` type usually corresponds to `llvm_int 1`, there are some
corner cases where it instead corresponds to `llvm_int 8`, which can cause your
specifications to fail if you are not careful. This has tripped up multiple
people in the past, and while it is difficult to catch this sort of mistake in
general, we can at least warn about it in the SAW manual, which this patch
accomplishes.