Support `mir_alloc`, `mir_alloc_mut`, and `mir_points_to`
This patch adds support for specifying the memory layouts of allocations in MIR
specifications through the following commands:

* `mir_alloc`, which allocates an immutable reference, and `mir_alloc_mut`,
  which allocates a mutable reference.
* `mir_points_to`, which declares that a reference points to a given value in a
  pre- or post-condition.

These commands behave much like `llvm_alloc` and `llvm_points_to`, and the code
for the MIR commands looks quite similar to the LLVM code as a result. The
largest difference implementation-wise is that I need to call into the
`crucible-mir` memory model in order to read values from and write values to
references. This patch therefore bumps the `crucible` submodule to bring in the
changes from https://github.com/GaloisInc/crucible/issues/1105, which make it
much more straightforward to use the `crucible-mir` memory model in a SAW
setting.

Along the way, this patch also performs some mild refactoring:

* This adds a `ConditionMetadata` field to `MirPointsTo` for the benefit of
  using it in `learnPointsTo`. This required moving around some
  `ConditionMetadata` values in `crucible-mir-comp` as a result.
* This moves the `CheckPointsToType` data type from
  `SAWScript.Crucible.LLVM.Builtins` to
  `SAWScript.Crucible.Common.Setup.Builtins`. This is because
  `CheckPointsToType` is now used in the `saw-remote-api` code that powers
  `mir_points_to` (see `SAWServer.MIRCrucibleSetup`), and it doesn't make much
  sense for this code to import LLVM-specific code.

This checks off one box in #1859.