mir_tuple_value
This adds support for a `mir_tuple_value` function in the MIR backend, which
allows constructing tuple `SetupValue`s. This is distinct from the notion of
structs in MIR, so there are now separate `SetupStruct` and `SetupTuple`
constructors for `SetupValue` to encode this distinction. At the moment, only
the MIR backend makes use of `SetupTuple`s, and they are not used at all in the
LLVM or JVM backends.

In the SAW remote API, the `"tuple"` type of `setup value` previously referred
to `SetupStruct`s, but this becomes especially confusing when SAW supports both
structs _and_ tuples. There are now separate `"struct"` and `"tuple"` `setup
value`s to properly tell them apart.

This checks off one box in #1859.