Handle `TError` in constrant guard translator
When using constraint guards in instantiations of parameterized modules,
Cryptol may replace impossible constraints with `TError`. To support
this behavior, this change adds a case to
`importNumericConstraintAsBool` that translates those `TError` constraints
to `False`. This is consistent with
[`checkProp` in Cryptol](https://github.com/GaloisInc/cryptol/blob/3973b15236ace5c11fdfabbf811d97daee7885ed/src/Cryptol/Eval.hs#L247),
which is the Cryptol function that handles evaluation of constraint
guard constraints.