Add composition support to `prove_bisim`
This change enables the reuse of proofs in `prove_bisim`.  `prove_bisim`
now returns a `BisimTheorem` that can be passed in to future
`prove_bisim` commands to facilitate compositional proofs.  To support
this, `prove_bisim` now also requires an additional relation over
states.  For a detailed description of how this all works, see the
module level doc comment in `SAWScript.Bisimulation`.