Start a SAW Rust verification tutorial
This patch adds an incomplete tutorial on how to use SAW to verify Rust code,
which complements the existing tutorial that covers LLVM and JVM verification.
This tutorial currently covers everything that has been implemented in SAW,
which includes:

* `mir-json`
* `mir_verify` basics
* References
* Arrays
* Tuples
* Structs
* Enums
* `mir_fresh_expanded_value`
* Slices
* Overrides
* Statics

The final section of the tutorial describes a case study using SAW to verify a
real-world piece of Rust code that implements the Salsa20 stream cipher.

Related to #1859.