Add declare_ghost_state, mark llvm_declare_ghost_state as legacy command
We want to be able to declare ghost state in all SAW language backends, so we
now use a more general `declare_ghost_state` name for the command to reflect
this. The old `llvm_declare_ghost_state` command is now a legacy command whose
use is discouraged in favor of `declare_ghost_state`.

This is one part of an eventual fix for #1929.