CI: Remove `retry cabal build` logic
The logic for retrying the `cabal build` command three consecutive times in a
row upon failure was apparently added to work around macOS-related dylib
issues, but it is unclear (and unlikely) if these issues still exist. Moreover,
this logic has the distinct disadvantage of potentially masking more serious
issues, such as the issues observed in #1961.

In principle, `cabal build` should not be an especially flaky part of the CI—it
should either work 100% of the time or fail 100% of the time. Let's just remove
the `retry cabal build` logic to make it easier to notice issues with `cabal
build` in the future. We can always revisit this choice later if need be.