Make symbolic branching work with JVM/MIR ghost state
This is not entirely straightforward to do for the JVM backend, as the
`jvmSimContext` function (`crucible-jvm`'s API function for constructing a
`SimContext`) does not support adding additional intrinsic types, including the
intrinsic type needed to support SAW ghost state. For the time being, I have
simply inlined the implementation of `jvmSimContext` as needed, and I have
opened https://github.com/GaloisInc/crucible/issues/1128 as a reminder to
generalize `jvmSimContext` upstream.