Skip test_comp_bisim test with coverage enabled
Sadly, building SAW with coverage enabled causes it to use noticeably more
memory at runtime, which is enough to exceed the GitHub Actions CI runners'
maximum memory and cause the job to be SIGKILL'ed when running the
`test_comp_bisim` job. It's not obvious how to fix this, so we simply disable
the `test_comp_bisim` job in CI when running with coverage enabled.