Adjust how type errors are printed.
- Adjust the indentation/formatting to make it clear that each error
message is actually one and not two separate ones.

- Instead of always printing the outermost enclosing types and
sometimes the innermost mismatched types, print each recursive layer
of the mismatched types. This helps a lot if the type is large.

- Avoid treating each successive argument of functions (since they're
curried) as another layer to print. These prints don't help.

- Strengthen the way we carry around the errors while recursing to make
these things possible.

- We don't need to carry around the identity of the enclosing context
("m") while recursing; we print it once at the end of each message,
and that should be sufficient now that it no longer looks like it's an
entirely separate message.

Printing multiple layers of mismatch was inspired by what the OCaml
compiler does with complicated types, but none of it's directly
analogous.

Note that the actual location of the error is printed by the caller.
We were printing the location of the enclosing context as if it was
the location of an additional error, and that was part of what was
significantly confusing; it's indented now in a way that makes it
clear it's supplemental information.

This change should be an improvement, but it's not meant to be (and
isn't) the definitive word on any of this. There are still several
things that can be improved:
   - when there's a mismatched function argument, it would be nice to
     identify which argument it is;
   - when there are mismatched record types, it would be nice to
     identify and print only the mismatched fields (this is especially
     significant because we have ad hoc records);
   - some of the callers definitely have the expected/found type
     arguments backwards;
   - the source positions (in the callers) are not as granular as one
     would like, e.g. an argument mismatch in a call is usually
     attributed to the entire application expression.