CI: Bump entree-specs commit to include fix for GaloisInc/entree-specs#5
This bumps the pinned `entree-specs` commit to bring in the changes from
GaloisInc/entree-specs#6, which requires `entree-specs` to build with
`coq-itree.5.2.*`. Doing so is necessary to fix the CI failure observed in
GaloisInc/entree-specs#5. Moreover, GaloisInc/entree-specs#6 adds upper version
bounds to `coq-itree` in `entree-specs`' `opam` file, minimizing the likelihood
of such a breakage happening again in the future.