Simplify the FailMGU type.
There's no need to distinguish between Nothing and Just [] when
tracking the last-seen pair of function-type expected/found errors, so
we can just use []. Note that while perhaps this ought to be a Maybe
(String, String) since it's always either empty or two strings, the
code is much tidier when it's a list.

Also rework the big comment describing it.

(And, I wrote originally that the initial message line gets indented
by two spaces when it's printed by the caller. I'm going to assume I
wasn't delusional when I wrote that, and I'm pretty sure nobody's
changed it in the meantime, but it doesn't happen with simple test
cases. So adjust the wording to waffle.)