Add empty lines between each expected/found pair in type errors.
This sets off the pairs some so they're easier to look at.

(It seems to be superior in general to also include empty lines around
the list of pairs too. Might sometime want to drop the empty lines if
there's only one pair, but that's substantially more complicated and
not worthwhile for the time being.)