Improve position tracking for expressions and patterns.
Add positions to all expression and pattern nodes. Kill off the LExpr
and LPattern expression and pattern cases that held just a position.
The idea had been to always use the first enclosing such position.
This doesn't provide accurate enough positions for type errors (in
general for a type error you always want the exact expression you're
complaining about, especially if you're printing spans of positions
and not just line numbers), and, worse, with this approach it's hard
to tell if any given position is the one you actually ought to be
using.

Fix up some of the position usage where prompted by the ensuing typing
changes.

Kill off the currentExprPos field of the local state monad in MGU.hs,
which was for collecting the enclosing LExpr and is no longer needed.

In pursuit of all this, add two ops to Position.hs:
   - leadingPos, which gets the empty (no span) position at the beginning
     of the position of something else, which can be used for deriving the
     position of implicit elements;
   - maxSpan', which is like maxSpan but works for (exactly) two Positioned
     objects of potentially heterogeneous type. This allows dropping a
     number of explicit calls to getPos in places where that's verbose and
     ugly.